Icons and confirmation link inserted (delete function)

+++ trunk/webapp/public/javascripts/shoppinglist.mod.shoppinglists.js
@@ -46,8 +46,9 @@ var Shoppinglist = Shoppinglist || {};
         for (i = 0, len = data.shoppinglists.length; i < len; i++) {
             current_item = data.shoppinglists[i];
             tmpHtml.add('<li>');
+            tmpHtml.add('<img src="images/icon_note.png" height="20" width="20" border="0">');
             tmpHtml.add('<a href="#" sid="' + current_item.shoppinglistId + '">' + current_item.name + '</a>');
-            tmpHtml.add('<a class="delete" href="controller_proxy.php?controller=deleteshoppinglist&amp;sid=' + current_item.shoppinglistId + '" sid="' + current_item.shoppinglistId + '">[delete]</a>');
+            tmpHtml.add('<a class="delete" onclick="javascript:return confirm(\'Are you sure?\')" href="controller_proxy.php?controller=deleteshoppinglist&amp;sid=' + current_item.shoppinglistId + '" sid="' + current_item.shoppinglistId + '">[delete]</a>');
             tmpHtml.add('</li>');
         }
         tmpHtml.add('<ul>');
